animation of the header and footer
