+++ source/index.html
@@ -7,8 +7,7 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mysterious Vacation</title>
 </head>
-<body>
-
+<body id="body">
   <header class="page-header js-header">
     <div class="page-header__wrapper">
       <a href="#" class="page-header__logo" aria-label="Перейти на страницу FJORD.Inc">
@@ -82,7 +81,30 @@
     <section class="screen screen--hidden screen--intro" id="top">
       <div class="screen__wrapper">
         <div class="intro">
-          <h1 class="intro__title">Таинственный отпуск</h1>
+          <h1 class="intro__title">
+            <span class="intro__title-word">
+              <span>Т</span>
+              <span>а</span>
+              <span>и</span>
+              <span>н</span>
+              <span>с</span>
+              <span>т</span>
+              <span>в</span>
+              <span>е</span>
+              <span>н</span>
+              <span>н</span>
+              <span>ы</span>
+              <span>й</span>
+            </span>
+            <span class="intro__title-word">
+              <span>о</span>
+              <span>т</span>
+              <span>п</span>
+              <span>у</span>
+              <span>с</span>
+              <span>к</span>
+            <!-- </span> -->
+          </h1>
           <div class="intro__content">
             <div class="intro__info">
               <p class="intro__label">Даты проведения конкурса:</p>
