+++ source/js/script.js
@@ -8,7 +8,7 @@ import result from './modules/result.js';
 import form from './modules/form.js';
 import social from './modules/social.js';
 import FullPageScroll from './modules/full-page-scroll';
-import intro from './modules/intro';
+import mainScreen from './modules/main-screen';
 
 // init modules
 mobileHeight();
@@ -19,7 +19,10 @@ chat();
 result();
 form();
 social();
-intro();
 
 const fullPageScroll = new FullPageScroll();
 fullPageScroll.init();
+
+window.addEventListener(`load`, () => {
+  mainScreen();
+});
