+++ source/scss/blocks/page-header.scss
@@ -20,6 +20,7 @@
 }
 
 .page-header__wrapper {
+  // opacity: 0;
   @include container;
   display: flex;
   align-items: center;
@@ -35,6 +36,9 @@
   padding: 1rem 0;
   width: 10rem;
   color: $c-light;
+  opacity: 0;
+  transform: translateX(-40px);
+  transition: all 300ms ease;
 
   @media (min-width: $stop-scaling) {
     top: -5px;
@@ -171,6 +175,7 @@
 }
 
 .page-header__menu {
+  opacity: 0;
   font-family: $font-alt;
   text-transform: uppercase;
   font-weight: 400;
@@ -266,7 +271,7 @@
 
     &.active {
       &::after {
-        width: 100%;
+        width: 0;
       }
     }
 
@@ -306,5 +311,18 @@
   @media (max-width: $mobile) and (orientation: landscape) {
     margin-left: 0;
     width: 100%;
+  }
+}
+
+.is-load .page-header__menu {
+  opacity: 1;
+}
+
+.is-load .js-menu-link {
+  &.active {
+    &::after {
+      left: 0;
+      width: 100%;
+    }
   }
 }
