+++ source/scss/blocks/screen.scss
@@ -14,7 +14,7 @@
     min-height: calc(var(--vh, 1vh) * 100);
   }
 
-  @media (max-width: $tablet - 1) and (orientation: landscape) {
+  @media (max-width: $tablet) and (orientation: landscape) {
     height: 100vh;
   }
 }
@@ -46,6 +46,8 @@
   color: $c-dark;
   background-color: $c-light;
   z-index: 1;
+  opacity: 0;
+  transform: translateY(50px);
 
   @media (max-width: $tablet) and (orientation: portrait) {
     width: 3.5rem;
@@ -353,4 +355,9 @@
 
 .screen--rules {
   background-color: $c-dark;
+}
+
+.is-load .screen__footer {
+  opacity: 1;
+  transform: translateY(0);
 }
