+++ source/scss/blocks/social-block.scss
@@ -23,6 +23,7 @@
   height: 4rem;
   color: $c-dark;
   border-radius: 50%;
+  opacity: 0;
 
   &:hover {
     color: $c-purple;
@@ -39,12 +40,16 @@
   &::before {
     @include center;
     content: "";
+    position: absolute;
+    left: 0;
+    top: 0;
     width: 100%;
     height: 100%;
     border-radius: 50%;
     background-color: $c-light;
     border: 2px solid $c-light;
     z-index: -1;
+    transform: scale(0.5);
   }
 
   svg {
@@ -59,6 +64,9 @@
     color: $c-purple;
 
     &::before {
+      position: absolute;
+      left: -3px;
+      top: -3px;
       width: 112.5%;
       height: 112.5%;
     }
@@ -83,6 +91,8 @@
     color: $c-light;
 
     &::before {
+      left: 0;
+      top: 0;
       width: 100%;
       height: 100%;
       background-color: $c-purple;
@@ -252,5 +262,13 @@
         height: 10px;
       }
     }
+  }
+}
+
+.is-load .social-block__toggler {
+  opacity: 1;
+
+  &::before {
+    transform: scale(1);
   }
 }
