+++ source/scss/general/animation.scss
@@ -1,4 +1,6 @@
 .page-header__menu {
+  transition: all 0.4s cubic-bezier(0.08, 0.77, 0.36, 0.71) 200ms;
+
   & a {
     &::after {
       transition: width 0.2s ease;
@@ -10,6 +12,22 @@
       transition: width 0.2s ease;
     }
   }
+}
+
+.js-menu-link {
+  &.active {
+    &::after {
+      transition: width 0.2s ease 300ms;
+    }
+  }
+}
+
+.page-header__logo {
+  transition: all 200ms cubic-bezier(0.08, 0.77, 0.36, 0.71) 100ms;
+}
+
+.screen__footer {
+  transition: all 300ms cubic-bezier(0.08, 0.77, 0.36, 0.71);
 }
 
 .social-block__list li {
@@ -28,6 +46,85 @@
   transition-delay: 300ms;
 }
 
+.intro__title {
+  transition: all 100ms ease-out 400ms;
+}
+
+.intro__label {
+  transition: all 300ms ease-in-out 1.3s;
+}
+
+.intro__date {
+  transition: all 300ms ease 1.2s;
+}
+
+.intro__title-word {
+  & span {
+    transition: all 500ms ease 400ms;
+
+    &:nth-child(1) {
+      transition-delay: 450ms;
+    }
+
+    &:nth-child(2),
+    &:nth-child(4),
+    &:nth-child(6) {
+      transition-delay: 350ms;
+    }
+
+    &:nth-child(3),
+    &:nth-child(7),
+    &:nth-child(10),
+    &:nth-child(12) {
+      transition-delay: 100ms;
+    }
+
+    &:nth-child(5),
+    &:nth-child(9),
+    &:nth-child(11) {
+      transition-delay: 470ms;
+    }
+
+    &:nth-child(8) {
+      transition-delay: 500ms;
+    }
+  }
+}
+
+.intro__title-word:last-child {
+  transition: all 500ms ease 700ms;
+}
+
+.slider__item-title {
+  & span {
+    transition: all 300ms linear;
+
+    &:nth-child(1) {
+      transition-delay: 50ms;
+    }
+
+    &:nth-child(5) {
+      transition-delay: 90ms;
+    }
+
+    &:nth-child(even) {
+      transition-delay: 280ms;
+    }
+
+    &:nth-child(odd) {
+      transition-delay: 100ms;
+    }
+  }
+}
+
+.social-block__toggler {
+  transition: opacity 300ms cubic-bezier(0.08, 0.77, 0.42, 0.66) 300ms;
+
+  &::before {
+    transition: transform 300ms cubic-bezier(0.08, 0.77, 0.42, 0.66) 300ms;
+  }
+}
+
 .active {
   & .swiper-slide-active {
     & .slider__item-text {
@@ -43,26 +140,6 @@
 
   & .slider__item-title {
     animation: move-title 300ms 1 forwards linear;
-
-    & span {
-      transform: translateY(0);
-
-      &:nth-child(1) {
-        transition-delay: 50ms;
-      }
-
-      &:nth-child(5) {
-        transition-delay: 90ms;
-      }
-
-      &:nth-child(even) {
-        transition-delay: 280ms;
-      }
-
-      &:nth-child(odd) {
-        transition-delay: 100ms;
-      }
-    }
   }
 }
