+++ source/scss/general/key-frames.scss
@@ -17,7 +17,7 @@
   }
 
   to {
-    transform: translateX(0);
+    transform: translateY(0);
     opacity: 1;
   }
 }
